test base class and individual test classes
