+++ Web/Edubase.Web.UIUnitTests/Areas/Governors/Models/ReplaceChairViewModelTests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Edubase.Web.UIUnitTests.Areas.Governors.Models
+{
+    internal class ReplaceChairViewModelTests
+    {
+    }
+}
